fix: resolve API route merge conflict
